+++ samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/author/ItemModifyListener.java
@@ -438,7 +438,7 @@ public class ItemModifyListener implements ActionListener
 		}
 		// making sure if there is any partial credit in place we account for that --mustansar
 		if  (Long.valueOf(itemauthorbean.getItemType()).equals(TypeFacade.MULTIPLE_CHOICE)) {
-			answerbean.setPartialCredit(answerArray[i].getPartialCredit()); 
+			answerbean.setPartialCredit(answerArray[i].getPartialCredit().intValue()); 
 		}
 		answerbeanlist.add(answerbean);
          }
